Reorganization and prepping for PR
