+++ Assets/Prefabs/Ui/UI Canvases/InGameCanvasMaster.prefab
@@ -2480,11 +2480,11 @@ RectTransform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 6608078730428679260}
   - {fileID: 1467200242624961123}
   - {fileID: 9045590869081106698}
   - {fileID: 1254141111128829534}
   - {fileID: 2657118438954276819}
+  - {fileID: 6608078730428679260}
   - {fileID: 2890816136378274803}
   - {fileID: 8668228369754994297}
   - {fileID: 7123565336505001915}
