--- Assets/Resources/Effect Cards.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6f72f2e4d383f63468347d9873836f23
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
