+++ Assets/Scripts/Bartering Minigame/OfferedItemsEffects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d00a96dab1342cc41b3b059c2815abf8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
